Add `color` field to Mooblooms and Mooshrooms, alongside particle and nectar tweaks.

+++ Common/src/main/java/net/merchantpug/bovinesandbuttercups/client/particle/ModelLocationParticle.java
@@ -36,19 +36,19 @@ public class ModelLocationParticle extends TextureSheetParticle {
     }
 
     protected float getU0() {
-        return this.sprite.getU((double)((this.uo + 1.0F) / 4.0F * 16.0F));
+        return this.sprite.getU((this.uo + 1.0F) / 4.0F * 16.0F);
     }
 
     protected float getU1() {
-        return this.sprite.getU((double)(this.uo / 4.0F * 16.0F));
+        return this.sprite.getU(this.uo / 4.0F * 16.0F);
     }
 
     protected float getV0() {
-        return this.sprite.getV((double)(this.vo / 4.0F * 16.0F));
+        return this.sprite.getV(this.vo / 4.0F * 16.0F);
     }
 
     protected float getV1() {
-        return this.sprite.getV((double)((this.vo + 1.0F) / 4.0F * 16.0F));
+        return this.sprite.getV((this.vo + 1.0F) / 4.0F * 16.0F);
     }
 
     public int getLightColor(float p_108291_) {
